requirements.txt
aiohttp==3.6.2
aiohttp-retry==1.0
asgiref==3.2.7
async-timeout==3.0.1
asynctest==0.13.0
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.6
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
multidict==4.7.5
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.9
yarl==1.4.2
zipp==3.1.0
